[desk-tool] Fix pane footer’s z-offset (#2326)
* feat(core): add --mock-browser-env flag to `sanity exec` command (#2319)

* feat(cli): load `.env` files before running any command (#2320)

* feat(base): add z-offset values for `paneHeader` and `paneFooter`

* fix(desk-tool): improve z-offsets

* Make sure the pane header and footer are always on top of pane content

* feat(base): add `fullscreen` to z-offset values

* fix(desk-tool): use `fullscreen` z-offset for InspectDialog

* refactor(base): improve organization of z-offset API
