+++ packages/@sanity/base/src/components/zOffsets/ZIndexContext.ts
@@ -1,0 +1,10 @@
+import {createContext} from 'react'
+import {ZIndexContextValue} from './types'
+import {defaults} from './defaults'
+
+/**
+ * @todo: Rename to `ZOffsetsContext`
+ *
+ * @internal
+ */
+export const ZIndexContext = createContext<ZIndexContextValue>(defaults)
