+++ packages/@sanity/base/src/components/zOffsets/useZIndex.ts
@@ -1,0 +1,12 @@
+import {useContext} from 'react'
+import {ZIndexContextValue} from './types'
+import {ZIndexContext} from './ZIndexContext'
+
+/**
+ * @todo: Rename to `useZOffsets`
+ *
+ * @internal
+ */
+export function useZIndex(): ZIndexContextValue {
+  return useContext(ZIndexContext)
+}
